% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.data.frame.VPRM_driver_data}
\alias{as.data.frame.VPRM_driver_data}
\title{coerce to a data frame}
\usage{
\method{as.data.frame}{VPRM_driver_data}(x, ...)
}
\arguments{
\item{x}{VPRM_driver_data object}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
data frame with fields variables NEE_obs, T, PAR, rho_nir,
rho_swir, EVI, and phen.
}
\description{
method for VPRM_driver_data
}
\details{
The data field of the VPRM_driver_data object is kept.  Other
fields (short name, long name, etc.) are discarded.
}
\examples{
data(Park_Falls)
pfa_dd <- VPRM_driver_data(name_long="Park Falls",
                           name_short = "US-PFa",
                           lat=45.9459,
                           lon=-90.2723,
                           PFT='MF',
                           tower_date=PFa_tower_obs[['date']],
                           NEE_obs=PFa_tower_obs[['FC']],
                           T=PFa_tower_obs[['TA']],
                           PAR=PFa_tower_obs[['PAR']],
                           date_nir = PFa_refl[['date']],
                           rho_nir=PFa_refl[['nir']],
                           date_swir = PFa_refl[['date']],
                           rho_swir = PFa_refl[['swir']],
                           date_EVI = PFa_evi[['date']],
                           EVI=PFa_evi[['evi']],
                           phen=NA)
print(head(as.data.frame(pfa_dd)))
}
\author{
Timothy W. Hilton
}

