% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getLSWI}
\alias{getLSWI}
\title{calculate LSWI}
\usage{
getLSWI(rho_nir, rho_swir)
}
\arguments{
\item{rho_nir}{1xN numeric vector; near-infrared (841-876 nm) band
satellite-derived reflectance}

\item{rho_swir}{1xN numeric vector; 1628-1652 nm band
satellite-derived reflectance}
}
\value{
LSWI: numeric vector same size as rho_nir and rho_swir
containing LSWI values.
}
\description{
calculates land surface water index (LSWI) according to eqn 3 in
Mahadevan et al, 2007.
}
\details{
LSWI may not exceed 1.0 and may take negative values (Xiao et al 2004)
}
\examples{
data(Park_Falls)
pfa_dd <- VPRM_driver_data(name_long="Park Falls",
                           name_short = "US-PFa",
                           lat=45.9459,
                           lon=-90.2723,
                           PFT='MF',
                           tower_date=PFa_tower_obs[['date']],
                           NEE_obs=PFa_tower_obs[['FC']],
                           T=PFa_tower_obs[['TA']],
                           PAR=PFa_tower_obs[['PAR']],
                           date_nir = PFa_refl[['date']],
                           rho_nir=PFa_refl[['nir']],
                           date_swir = PFa_refl[['date']],
                           rho_swir = PFa_refl[['swir']],
                           date_EVI = PFa_evi[['date']],
                           EVI=PFa_evi[['evi']],
                           phen=NA)
pfa_df <- as.data.frame( pfa_dd )
LSWI <- getLSWI( rho_nir=pfa_df[['rho_nir']],
                 rho_swir=pfa_df[['rho_swir']] )
}
\author{
Timothy W. Hilton
}
\references{
Xiangming Xiao, Qingyuan Zhang, Bobby Braswell, Shawn
Urbanski, Stephen Boles, Steven Wofsy, Berrien Moore III, Dennis
Ojima, Modeling gross primary production of temperate deciduous
broadleaf forest using satellite images and climate data, Remote
Sensing of Environment, Volume 91, Issue 2, 30 May 2004, Pages
256-270, ISSN 0034-4257, 10.1016/j.rse.2004.03.010.

Mahadevan, P., Wofsy, S., Matross, D., Xiao, X., Dunn,
A., Lin, J., Gerbig, C., Munger, J., Chow, V., and Gottlieb, E.: A
satellite-based biosphere parameterization for net ecosystem CO2
exchange: Vegetation Photosynthesis and Respiration Model
(VPRM), Global Biogeochem. Cy., 22, GB2005,
doi:10.1029/2006GB002735, 2008.
}

