% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vprm.R
\name{getPscale_urban}
\alias{getPscale_urban}
\title{calculate Pscale for urbanVPRM}
\usage{
getPscale_urban(EVI)
}
\arguments{
\item{EVI}{1xN numeric vector; enhanced vegetation index}
}
\value{
Pscale term in urbanVPRM GEE equation (eqn SI eqn 4 in Hardiman et al, 2017)
}
\description{
calculates Pscale according to SI eqn 4 in Hardiman et al, 2017
}
\examples{
data(Park_Falls)
pfa_dd <- VPRM_driver_data(name_long="Park Falls",
                           name_short = "US-PFa",
                           lat=45.9459,
                           lon=-90.2723,
                           PFT='MF',
                           tower_date=PFa_tower_obs[['date']],
                           NEE_obs=PFa_tower_obs[['FC']],
                           T=PFa_tower_obs[['TA']],
                           PAR=PFa_tower_obs[['PAR']],
                           date_nir = PFa_refl[['date']],
                           rho_nir=PFa_refl[['nir']],
                           date_swir = PFa_refl[['date']],
                           rho_swir = PFa_refl[['swir']],
                           date_EVI = PFa_evi[['date']],
                           EVI=PFa_evi[['evi']],
                           phen=NA)
pfa_df <- as.data.frame(pfa_dd)
pscale <- getPscale_urban(pfa_df[['EVI']])
}
\references{
Hardiman, B. S., Wang, J. A., Hutyra, L. R., Gately, C. K.,
  Getson, J. M., & Friedl, M. A. (2017). Accounting for urban biogenic
  fluxes in regional carbon budgets. Science of The Total Environment, 592,
  366–372. https://doi.org/10.1016/j.scitotenv.2017.03.028
}
\author{
Timothy W. Hilton
}
