% Generated by roxygen2 (4.0.2): do not edit by hand
\name{interpMODIS}
\alias{interpMODIS}
\title{interpolate MODIS band data to provide values at a given set of dates}
\usage{
interpMODIS(mod_date, mod_val, out_date, method)
}
\arguments{
\item{mod_date}{chron vector; MODIS Time stamp}

\item{mod_val}{MODIS values}

\item{out_date}{chron vector; the dates for the output}

\item{method}{string, either "linear" or "constant"}
}
\value{
list with components "date" and "val", containing
timestamps (class chron) and values (class is dependent on the
MODIS product) of the interpolated data.
}
\description{
interpolate MODIS band data to provide values at a given set of
dates using either linear or constant interpolation.
}
\examples{
data(Park_Falls)
interpolated_EVI <- interpMODIS(mod_date=PFa_evi[['date']],
                                mod_val=PFa_evi[['evi']],
                                out_date=PFa_tower_obs[['date']],
                                method='linear')
plot(interpolated_EVI[['date']], interpolated_EVI[['val']],
     type='l', lty='dashed',
     xlab='date', ylab='EVI', main='Park Falls')
with(PFa_evi, points(date, evi, pch='*', cex=2.0, col='red'))
legend( x='topright',
       legend=c('observed', 'interpolated'),
       pch=c('*', NULL),
       col=c('red', 'black'),
       lty=c('blank', 'dashed'),
       pt.cex=c(2.0, 1.0))
}
\author{
Timothy W. Hilton
}

