% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateVPRMPars.R
\name{L}
\alias{L}
\title{cost function for VPRM parameterization}
\usage{
L(par, driver_data, model_form = "Mahadevan07")
}
\arguments{
\item{par}{four-element numeric vector containing lambda, alpha,
beta, PAR_0 (in that order)}

\item{driver_data}{May be a VPRM_driver_data object or a data
frame.  If a data frame, driver_data must contain the variables
Tscale, Pscale, Wscale, EVI, and PAR, and NEE_obs.}

\item{model_form}{string, optional; form of VPRM model to use. Options are
"Mahadevan07" (default) to use the VPRM formulation of Mahadevan et al.
(2007), or "urban" to use the urbanVPRM formulation of Hardiman et al.
(2017). If set to "urban", the driver data must include variables ISA
proportion (impervious surface area, 0.0 to 1.0) and refEVI (reference
EVI).}
}
\value{
sum of squared errors for VPRM NEE residuals for the
specified parameters and NEE observations.
}
\description{
calculate sum of squared errors (SSE) for VPRM NEE, given a set of VPRM
parameters and a a set of NEE observations.
}
\details{
"L" abbreviations likelihood; currently SSE is a standin for a statistically proper likelihood function for land surface model residuals.
}
\references{
Mahadevan, P., Wofsy, S., Matross, D., Xiao, X., Dunn, A., Lin,
  J., Gerbig, C., Munger, J., Chow, V., and Gottlieb, E.: A satellite-based
  biosphere parameterization for net ecosystem CO2 exchange: Vegetation
  Photosynthesis and Respiration Model (VPRM), Global Biogeochem. Cy., 22,
  GB2005, doi:10.1029/2006GB002735, 2008.

Hardiman, B. S., Wang, J. A., Hutyra, L. R., Gately, C. K.,
  Getson, J. M., & Friedl, M. A. (2017). Accounting for urban biogenic
  fluxes in regional carbon budgets. Science of The Total Environment, 592,
  366–372. https://doi.org/10.1016/j.scitotenv.2017.03.028
}
\author{
Timothy W. Hilton
}
