% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateVPRMPars.R
\name{optimizeVPRM_DE}
\alias{optimizeVPRM_DE}
\title{run DEoptim to estimate VPRM parameter values}
\usage{
optimizeVPRM_DE(
  driver_data,
  DEcontrol,
  refresh = 5,
  msg = "",
  lambda_prior = c(0, 1.5),
  alpha_prior = c(0, 1.5),
  beta_prior = c(-4, 4),
  PAR0_prior = c(0.1, 6000),
  model_form = "Mahadevan07"
)
}
\arguments{
\item{driver_data}{May be a VPRM_driver_data object or a data frame. If a
data frame, driver_data must contain the variables Tscale, Pscale, Wscale,
EVI, and PAR, and NEE_obs.}

\item{DEcontrol}{DEoptim.control object}

\item{refresh}{how many DE iterations in between progress writeouts}

\item{msg}{character string; message to be displayed before DE is called.
Useful for log files.}

\item{lambda_prior}{2-element vector; upper and lower lambda values for
optimization. Default is c(0.0, 1.5).}

\item{alpha_prior}{2-element vector; upper and lower alpha values for
optimization. Default is c(0.0, 1.5).}

\item{beta_prior}{2-element vector; upper and lower beta values for
optimization. Default is c(-4.0, 4.0).}

\item{PAR0_prior}{2-element vector; upper and lower PAR_0 values for
optimization. Default is c(0.1, 6000).}

\item{model_form}{string, optional; form of VPRM model to use. Options are
"Mahadevan07" (default) to use the VPRM formulation of Mahadevan et al.
(2007), or "urban" to use the urbanVPRM formulation of Hardiman et al.
(2017). If set to "urban", the driver data must include variables ISA
proportion (impervious surface area, 0.0 to 1.0) and refEVI (reference
EVI).}
}
\value{
DEoptim object. DEoptim output for best VPRM parameter estimate.
}
\description{
define VPRM parameter windows (upper and lower) and call DEoptim
to estimate VPRM parmeters.  Helper function for
estimate_VPRM_pars.
}
\examples{
data(Park_Falls)
}
\references{
Mahadevan, P., Wofsy, S., Matross, D., Xiao, X., Dunn, A., Lin,
  J., Gerbig, C., Munger, J., Chow, V., and Gottlieb, E.: A satellite-based
  biosphere parameterization for net ecosystem CO2 exchange: Vegetation
  Photosynthesis and Respiration Model (VPRM), Global Biogeochem. Cy., 22,
  GB2005, doi:10.1029/2006GB002735, 2008.

Hardiman, B. S., Wang, J. A., Hutyra, L. R., Gately, C. K.,
  Getson, J. M., & Friedl, M. A. (2017). Accounting for urban biogenic
  fluxes in regional carbon budgets. Science of The Total Environment, 592,
  366–372. https://doi.org/10.1016/j.scitotenv.2017.03.028
}
\author{
Timothy W. Hilton
}
