% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPRM_driver_data.R
\name{SWDN_2_PAR}
\alias{SWDN_2_PAR}
\title{downward shortwave radiation to PAR}
\usage{
SWDN_2_PAR(SWDN)
}
\arguments{
\item{SWDN}{numeric vector; downward shortwave radiation (W m-2)}
}
\value{
numeric vector containing PAR values (umol m-2 s-1)
}
\description{
calculate photosynthetically available radiation (PAR) from incident
downward shortwave radiation (SWDN)
}
\details{
uses the mean result from Britton and Dodd (1976) table I. Averages the
tables two values for Apr-May together, then averages all bimonthly
averages.
}
\references{
Britton, C. M. and Dodd, J. D., 1976. Relationships of
  photosynthetically active radiation and shortwave irradiance. Agric.
  Meteorol., 17: 1--7. https://doi.org/10.1016/0002-1571(76)90080-7
}
\author{
Timothy W. Hilton
}
