% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vprm.R
\name{urbanvprm_calc_R}
\alias{urbanvprm_calc_R}
\title{calculate urbanVPRM ecosystem respiration}
\usage{
urbanvprm_calc_R(Rinit, driver_data, alpha_param, beta_param)
}
\arguments{
\item{Rinit}{numeric; the Rinit term, equal to the respiration for the
Mahadevan et al (2007) formulation.}

\item{driver_data}{May be a VPRM_driver_data object or a data frame. If a
data frame, driver_data must contain the variables Tscale, Pscale, Wscale,
EVI, PAR, and Tresp. The variables alpha and beta are optional (see
'details').}

\item{alpha_param}{numeric, optional; VPRM parameter (slope of respiration
with respect to temperature)}

\item{beta_param}{numeric, optional; VPRM parameter (basal respiration rate)}

\item{model_form}{string, optional; form of VPRM model to use. Options are
"Mahadevan07" (default) to use the VPRM formulation of Mahadevan et al.
(2007), or "urban" to use the urbanVPRM formulation of Hardiman et al.
(2017). If set to "urban", the driver data must include variables ISA
proportion (impervious surface area, 0.0 to 1.0) and refEVI (reference
EVI).}
}
\value{
vector of same length as number of rows in driver_data containin
  urbanVPRM ecosystem respiration [umol m-2 s-1]
}
\description{
calculates urbanVPRM ecosystem respiration (R) according to Hardiman et al.
(2017) SI eqn 7 and SI eqn 8 as described in SI section S2.4.
}
\details{
.. content for \details{} ..
}
\references{
Hilton, T. W., Davis, K. J., Keller, K., and Urban, N. M.:
  Improving North American terrestrial CO2 flux diagnosis using spatial
  structure in land surface model residuals, Biogeosciences, 10, 4607-4625,
  doi:10.5194/bg-10-4607-2013, 2013.

Mahadevan, P., Wofsy, S., Matross, D., Xiao, X., Dunn, A., Lin,
  J., Gerbig, C., Munger, J., Chow, V., and Gottlieb, E.: A satellite-based
  biosphere parameterization for net ecosystem CO2 exchange: Vegetation
  Photosynthesis and Respiration Model (VPRM), Global Biogeochem. Cy., 22,
  GB2005, doi:10.1029/2006GB002735, 2008.

Hardiman, B. S., Wang, J. A., Hutyra, L. R., Gately, C. K.,
  Getson, J. M., & Friedl, M. A. (2017). Accounting for urban biogenic
  fluxes in regional carbon budgets. Science of The Total Environment, 592,
  366–372. https://doi.org/10.1016/j.scitotenv.2017.03.028
}
\author{
Timothy W. Hilton
}
