% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vprm.R
\name{vprm_calc_GEE}
\alias{vprm_calc_GEE}
\title{calculate VPRM GEE}
\usage{
vprm_calc_GEE(driver_data, lambda_param = NULL, PAR_0_param = NULL)
}
\arguments{
\item{driver_data}{May be a VPRM_driver_data object or a data
frame.  If a data frame, driver_data must contain the variables
Tscale, Pscale, Wscale, EVI, and PAR.  The variables
lambda, and PAR_0 are optional (see 'details').}

\item{lambda_param}{numeric, optional; VPRM parameter: maximum
light use efficiency.}

\item{PAR_0_param}{numeric, optional; VPRM parameter (LUE
half-saturation value)}
}
\value{
vector of same length as number of rows in driver_data containin
VPRM GEE [umol m-2 s-1]
}
\description{
calculate VPRM GEE according to Mahadevan et al (2008) eq. 9
}
\details{
Arguments lambda and PAR_0 may be omitted from the function call.
In this case they must be present as variables in data frame
driver_data.  If either of these parameters are present in
driver_data *and* specified as parameters in the function call the
function parameter values will be used and the values in
driver_data will be ignored.  If specifed as function parameters
lambda and PAR_0 may be single values or numeric vectors the same
length as the number of observations in driver_data.
}
\examples{
data(Park_Falls)
pfa_dd <- VPRM_driver_data(name_long="Park Falls",
                           name_short = "US-PFa",
                           lat=45.9459,
                           lon=-90.2723,
                           PFT='MF',
                           tower_date=PFa_tower_obs[['date']],
                           NEE_obs=PFa_tower_obs[['FC']],
                           T=PFa_tower_obs[['TA']],
                           PAR=PFa_tower_obs[['PAR']],
                           date_nir = PFa_refl[['date']],
                           rho_nir=PFa_refl[['nir']],
                           date_swir = PFa_refl[['date']],
                           rho_swir = PFa_refl[['swir']],
                           date_EVI = PFa_evi[['date']],
                           EVI=PFa_evi[['evi']],
                           phen=NA)
data(VPRM_parameters)
attach(all_all_VPRM_parameters)
GEE <- vprm_calc_GEE(pfa_dd, lambda=lambda, PAR_0=PAR_0)
}
\references{
Mahadevan, P., Wofsy, S., Matross, D., Xiao, X., Dunn,
A., Lin, J., Gerbig, C., Munger, J., Chow, V., and Gottlieb, E.: A
satellite-based biosphere parameterization for net ecosystem CO2
exchange: Vegetation Photosynthesis and Respiration Model
(VPRM), Global Biogeochem. Cy., 22, GB2005,
doi:10.1029/2006GB002735, 2008.
}
\author{
Timothy W. Hilton
}
