% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vprm_calc_NEE}
\alias{vprm_calc_NEE}
\title{calculate VPRM NEE}
\usage{
vprm_calc_NEE(driver_data, lambda = NULL, alpha = NULL, beta = NULL,
  PAR_0 = NULL)
}
\arguments{
\item{driver_data}{May be a VPRM_driver_data object or a data
frame.  If a data frame, driver_data must contain the variables
Tscale, Pscale, Wscale, EVI, PAR, and Tresp.  The variables
lambda, alpha, beta, and PAR_0 are optional (see 'details').}

\item{lambda}{numeric, optional; VPRM parameter: maximum light use
efficiency.}

\item{alpha}{numeric, optional; VPRM parameter (slope of respiration with
respect to temperature)}

\item{beta}{numeric, optional; VPRM parameter (basal respiration rate)}

\item{PAR_0}{numeric, optional; VPRM parameter (LUE half-saturation value)}
}
\value{
vector of same length as number of rows in driver_data containin
VPRM NEE [umol m-2 s-1]
}
\description{
calculates VPRM net ecosystem exchange (NEE) according to eqn 12
in Mahadevan et al, 2007
}
\details{
Arguments lambda, alpha, beta, and PAR_0 may be omitted from the
function call.  In this case they must be present as variables in
data frame driver_data.  If any of these four parameters are present in
driver_data *and* specified as parameters in the function call the
function parameter values will be used and the values in driver_data
will be ignored.  If specifed as function parameters lambda,
alpha, beta, and PAR_0 may be single values or numeric vectors the
same length as the number of observations in driver_data.

The Tresp variable in driver_data is the temperature used to calculate
respiration.  Tresp should be max(Tair, Tlow), where Tair is the
air temperature (deg C) and Tlow is the minimum air temperature
(deg C) for respiration.  This is explained more fully in
Mahadevan et al (2008) section 2.2.
}
\examples{
data(Park_Falls)
pfa_dd <- VPRM_driver_data(name_long="Park Falls",
                           name_short = "US-PFa",
                           lat=45.9459,
                           lon=-90.2723,
                           PFT='MF',
                           tower_date=PFa_tower_obs[['date']],
                           NEE_obs=PFa_tower_obs[['FC']],
                           T=PFa_tower_obs[['TA']],
                           PAR=PFa_tower_obs[['PAR']],
                           date_nir = PFa_refl[['date']],
                           rho_nir=PFa_refl[['nir']],
                           date_swir = PFa_refl[['date']],
                           rho_swir = PFa_refl[['swir']],
                           date_EVI = PFa_evi[['date']],
                           EVI=PFa_evi[['evi']],
                           phen=NA)
data(VPRM_parameters)
attach(all_all_VPRM_parameters)
NEE <- vprm_calc_NEE(pfa_dd,
                     lambda=lambda, PAR_0=PAR_0, alpha=alpha, beta=beta)
}
\author{
Timothy W. Hilton
}
\references{
Hilton, T. W., Davis, K. J., Keller, K., and Urban,
N. M.: Improving North American terrestrial CO2 flux diagnosis
using spatial structure in land surface model residuals,
Biogeosciences, 10, 4607-4625, doi:10.5194/bg-10-4607-2013, 2013.

Mahadevan, P., Wofsy, S., Matross, D., Xiao, X., Dunn,
A., Lin, J., Gerbig, C., Munger, J., Chow, V., and Gottlieb, E.: A
satellite-based biosphere parameterization for net ecosystem CO2
exchange: Vegetation Photosynthesis and Respiration Model
(VPRM), Global Biogeochem. Cy., 22, GB2005,
doi:10.1029/2006GB002735, 2008.
}

