% Generated by roxygen2 (4.0.2): do not edit by hand
\name{VPRM_driver_data}
\alias{VPRM_driver_data}
\title{build a VPRM_driver_data object}
\usage{
VPRM_driver_data(name_long = "", name_short = "", lat = NA, lon = NA,
  PFT = NA, note = "", Tmin = 0, Tmax = 40, Topt = 20, Tlow = 2,
  tower_date = NA, NEE_obs = NA, T = NA, PAR = NA, date_nir = NA,
  rho_nir = NA, date_swir, rho_swir = NA, date_EVI, EVI = NA, phen = NA)
}
\arguments{
\item{name_long}{character string; "short name" of the site.  e.g. US-PFa}

\item{name_short}{character string; "long name" of the site.  e.g. Park Falls}

\item{lat}{numeric; latitude of the site (deg N)}

\item{lon}{numeric; longitude of the site (deg E)}

\item{PFT}{character string; plant functional type.  Will be converted to a factor.}

\item{note}{character string; optional note; could be anything the
user finds useful.}

\item{Tmin}{numeric; minimum temperature for photosynthesis (deg
C).  See Mahadevan et al (2008) eq. 6.}

\item{Tmax}{numeric; maximim temperature for photosynthesis (deg
C).  See Mahadevan et al (2008) eq. 6.}

\item{Topt}{numeric; optimum temperature for photosynthesis (deg
C).  See Mahadevan et al (2008) eq. 6.}

\item{Tlow}{numeric; minimum temperature for respiration (deg C).
See Mahadevan et al (2008) eq. 10.}

\item{tower_date}{chron vector; timestamps for all tower
observations (NEE_obs, T, PAR)}

\item{NEE_obs}{numeric vector; eddy covariance observed net
ecosystem exchange (NEE, umol m-2 s-1)}

\item{T}{numeric vector; observed air temperature (deg C)}

\item{PAR}{numeric vector; observed photosynthetically active
radiation (umol m-2 s-1)}

\item{date_nir}{chron vector; timestamps for NIR reflectance.}

\item{rho_nir}{numeric vector; NIR reflectance values}

\item{date_swir}{chron vector; timestamps for SWIR reflectance.}

\item{rho_swir}{numeric vector; SWIR reflectance values}

\item{date_EVI}{chron vector; timestamps for enhanced vegetation index (EVI).}

\item{EVI}{numeric vector; EVI values.}

\item{phen}{factor; phenology dynamics.  levels are ginc (onset
greenness increase), gdec (onset greenness decrease), gmin (onset
greenness minimum), gmax (onset greenness maximum).  If not
specified, phenology dynsamics are calculated from EVI using a
method similar to Zhang et al (2003).}
}
\value{
an object of class VPRM_driver_data.  Has fields (see
above arguments for definitions): name_long, name_short, lat, lon,
PFT, note, Tmin, Tmax, Topt, Tlow.  The data for NEE_obs, T, PAR,
rho_nir, rho_swir, EVI, and phen should be accessed using the
as.data.frame method.
}
\description{
class constructor for VPRM_driver_data.  Accepts all driver data
necessary to run VPRM for a single eddy covariance site,
calculates "derived" fields, and interpolates phenology dynamics.
}
\details{
"Derived fields" denote fields that are calculated from other
observed quanitities.  For example, land surface water index
(LSWI) is a derived field, as it is calculated from MODIS
reflectances in the short infrared and near infrared bands.
}
\examples{
data(Park_Falls)
pfa_dd <- VPRM_driver_data(name_long="Park Falls",
                           name_short = "US-PFa",
                           lat=45.9459,
                           lon=-90.2723,
                           PFT='MF',
                           tower_date=PFa_tower_obs[['date']],
                           NEE_obs=PFa_tower_obs[['FC']],
                           T=PFa_tower_obs[['TA']],
                           PAR=PFa_tower_obs[['PAR']],
                           date_nir = PFa_refl[['date']],
                           rho_nir=PFa_refl[['nir']],
                           date_swir = PFa_refl[['date']],
                           rho_swir = PFa_refl[['swir']],
                           date_EVI = PFa_evi[['date']],
                           EVI=PFa_evi[['evi']],
                           phen=NA)
print(head(as.data.frame(pfa_dd)))
}
\author{
Timothy W. Hilton
}
\references{
Mahadevan, P., S. C. Wofsy, D. M. Matross, X. Xiao,
A. L. Dunn, J. C. Lin, C. Gerbig, J. W. Munger, V. Y. Chow, and
E. W. Gottlieb (2008), A satellite-based biosphere
parameterization for net ecosystem CO2 exchange: Vegetation
Photosynthesis and Respiration Model (VPRM), Global
Biogeochem. Cycles, 22, GB2005, doi:10.1029/2006GB002735.

Xiaoyang Zhang, Mark A. Friedl, Crystal B. Schaaf,
Alan H. Strahler, John C.F. Hodges, Feng Gao, Bradley C. Reed,
Alfredo Huete, Monitoring vegetation phenology using MODIS, Remote
Sensing of Environment, Volume 84, Issue 3, March 2003, Pages
471-475, ISSN 0034-4257,
http://dx.doi.org/10.1016/S0034-4257(02)00135-9.
}

