% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPRMLandSfcModel-package.R
\docType{package}
\name{VPRMLandSfcModel-package}
\alias{VPRMLandSfcModel}
\alias{VPRMLandSfcModel-package}
\title{VPRMLandSfcModel: R implementation of VPRM with parameter estimation}
\description{
Provides an R implementation of the Vegetation Photosynthesis and Respiration Model (VPRM), the urbanVPRM, and tools to estimate VPRM parameter values from observations. VPRM is a simple diagnostic land surface model based on light-use efficiency. VPRM diagnoses gross ecosystem exchange (GEE) of carbon dioxide, ecosystem respiration (R), and net ecosystem exchange (NEE) of carbon dioxide. To cite the VPRMLandSfcModel R package or the (urban)VPRM model itself in a publication, or for a more detailed description of VPRM, see (within R) 'citation("VPRMLandSurfaceModel")'.
}
\keyword{internal}
