\name{VPRMModel-package}
\alias{VPRMModel-package}
\alias{VPRMModel}
\docType{package}
\title{
R implementation of VPRM
}
\description{
Provides an R implementation of Harvard University's Vegetation
Photosynthesis and Respiration Model (VPRM).  See references (below) for
a detailed description of VPRM.
}
\details{
\tabular{ll}{
Package: \tab VPRMModel\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-08-13\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
The function NEE calculates net ecosystem exchange (NEE) according to
the VPRM model.
}
\author{
Timothy W. Hilton

Maintainer: Timothy W. Hilton <hilton@meteo.psu.edu>
}
\references{
Mahadevan, P., Wofsy, S.C., Matross, D.M., Xiao, X., Dunn, A.L., Lin, J.C., Gerbig, C., Munger, J.W., Chow, V.Y., and Gottlieb, E.W., 2008. A satellite-based biosphere parameterization for net ecosystem CO2  exchange: Vegetation photosynthesis and respiration model (VPRM). Global Biogeochemical Cycles, 22, GB2005. doi:10.1029/2006GB002735.
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }

