% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vprm.R
\name{getPscale}
\alias{getPscale}
\title{calculate Pscale}
\usage{
getPscale(LSWI, phen)
}
\arguments{
\item{LSWI}{1xN numeric vector; land surface water index}

\item{phen}{1xN factor vector; MODIS phenology (factor with levels
ginc, gmin, gmax, gdec)}
}
\value{
Pscale term in VPRM equation (eqn 7 in Mahadevan et al, 2008)
}
\description{
calculates Pscale according to eqn 7 in Mahadevan et al, 2007
}
\details{
The levels of phen are abbreviations: ginc - "onset greenness
increase", gmin - "onset greenness minimum", gmax - "onset
greenness maximum", gdec - "onset greenness decrease".  This
phenology dynamics classification scheme is explained more fully
in Zhang et al (2003).
}
\examples{
data(Park_Falls)
pfa_dd <- VPRM_driver_data(name_long="Park Falls",
                           name_short = "US-PFa",
                           lat=45.9459,
                           lon=-90.2723,
                           PFT='MF',
                           tower_date=PFa_tower_obs[['date']],
                           NEE_obs=PFa_tower_obs[['FC']],
                           T=PFa_tower_obs[['TA']],
                           PAR=PFa_tower_obs[['PAR']],
                           date_nir = PFa_refl[['date']],
                           rho_nir=PFa_refl[['nir']],
                           date_swir = PFa_refl[['date']],
                           rho_swir = PFa_refl[['swir']],
                           date_EVI = PFa_evi[['date']],
                           EVI=PFa_evi[['evi']],
                           phen=NA)
pfa_df <- as.data.frame( pfa_dd )
pscale <- getPscale( pfa_df[['LSWI']], pfa_df[['phen']] )
}
\references{
Mahadevan, P., Wofsy, S., Matross, D., Xiao, X., Dunn,
A., Lin, J., Gerbig, C., Munger, J., Chow, V., and Gottlieb, E.: A
satellite-based biosphere parameterization for net ecosystem CO2
exchange: Vegetation Photosynthesis and Respiration Model
(VPRM), Global Biogeochem. Cy., 22, GB2005,
doi:10.1029/2006GB002735, 2008.

Xiaoyang Zhang, Mark A. Friedl, Crystal B. Schaaf,
Alan H. Strahler, John C.F. Hodges, Feng Gao, Bradley C. Reed,
Alfredo Huete, Monitoring vegetation phenology using MODIS, Remote
Sensing of Environment, Volume 84, Issue 3, March 2003, Pages
471-475, ISSN 0034-4257,
http://dx.doi.org/10.1016/S0034-4257(02)00135-9.
}
\author{
Timothy W. Hilton
}
