% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWscale}
\alias{getWscale}
\title{calculate Wscale}
\usage{
getWscale(LSWI, LSWI_max)
}
\arguments{
\item{LSWI}{numeric vector; land surface water index}

\item{LSWI_max}{numeric; maximum LSWI for site (that is, a single value)}
}
\value{
Pscale term in VPRM equation (eqn 12 in Mahadevan et al, 2007)
}
\description{
calculates Wscale according to eqn 8 in Mahadevan et al, 2007
}
\examples{
data(Park_Falls)
pfa_dd <- VPRM_driver_data(name_long="Park Falls",
                           name_short = "US-PFa",
                           lat=45.9459,
                           lon=-90.2723,
                           PFT='MF',
                           tower_date=PFa_tower_obs[['date']],
                           NEE_obs=PFa_tower_obs[['FC']],
                           T=PFa_tower_obs[['TA']],
                           PAR=PFa_tower_obs[['PAR']],
                           date_nir = PFa_refl[['date']],
                           rho_nir=PFa_refl[['nir']],
                           date_swir = PFa_refl[['date']],
                           rho_swir = PFa_refl[['swir']],
                           date_EVI = PFa_evi[['date']],
                           EVI=PFa_evi[['evi']],
                           phen=NA)
pfa_df <- as.data.frame( pfa_dd )
wscale <- getWscale(pfa_df[['LSWI']], max(pfa_df[['LSWI']], na.rm=TRUE ))
}
\author{
Timothy W. Hilton
}
\references{
Mahadevan, P., Wofsy, S., Matross, D., Xiao, X., Dunn,
A., Lin, J., Gerbig, C., Munger, J., Chow, V., and Gottlieb, E.: A
satellite-based biosphere parameterization for net ecosystem CO2
exchange: Vegetation Photosynthesis and Respiration Model
(VPRM), Global Biogeochem. Cy., 22, GB2005,
doi:10.1029/2006GB002735, 2008.
}

